remove unnecessary sets

+++ contracts/src/v0.8/keystone/test/CapabilityRegistry_RemoveNodeOperatorsTest.t.sol
@@ -5,10 +5,10 @@ import {BaseTest} from "./BaseTest.t.sol";
 import {CapabilityRegistry} from "../CapabilityRegistry.sol";
 
 contract CapabilityRegistry_RemoveNodeOperatorsTest is BaseTest {
-  event NodeOperatorRemoved(uint256 nodeOperatorId);
+  event NodeOperatorRemoved(uint32 nodeOperatorId);
 
   uint32 private constant TEST_NODE_OPERATOR_ONE_ID = 1;
-  uint256 private constant TEST_NODE_OPERATOR_TWO_ID = 2;
+  uint32 private constant TEST_NODE_OPERATOR_TWO_ID = 2;
 
   function setUp() public override {
     BaseTest.setUp();
@@ -19,7 +19,7 @@ contract CapabilityRegistry_RemoveNodeOperatorsTest is BaseTest {
   function test_RevertWhen_CalledByNonOwner() public {
     changePrank(STRANGER);
     vm.expectRevert("Only callable by owner");
-    uint256[] memory nodeOperatorsToRemove = new uint256[](2);
+    uint32[] memory nodeOperatorsToRemove = new uint32[](2);
     nodeOperatorsToRemove[1] = 1;
     s_capabilityRegistry.removeNodeOperators(nodeOperatorsToRemove);
   }
@@ -31,7 +31,7 @@ contract CapabilityRegistry_RemoveNodeOperatorsTest is BaseTest {
     emit NodeOperatorRemoved(TEST_NODE_OPERATOR_ONE_ID);
     vm.expectEmit(true, true, true, true, address(s_capabilityRegistry));
     emit NodeOperatorRemoved(TEST_NODE_OPERATOR_TWO_ID);
-    uint256[] memory nodeOperatorsToRemove = new uint256[](2);
+    uint32[] memory nodeOperatorsToRemove = new uint32[](2);
     nodeOperatorsToRemove[0] = TEST_NODE_OPERATOR_ONE_ID;
     nodeOperatorsToRemove[1] = TEST_NODE_OPERATOR_TWO_ID;
     s_capabilityRegistry.removeNodeOperators(nodeOperatorsToRemove);
